feat(api): add auth blueprint
